Update icon dropdowns and pending trx
