+++ public/locales/en/translation.json
@@ -100,6 +100,7 @@
         "notEnoughGas": "Not enough ETH for the gas",
         "noAssets": "There are no assets",
         "noTransaction": "There are no transactions",
+        "noRecentTransaction": "Your transactions will appear here",
         "noTransactionDesc": "Please make some transactions first.",
         "noPlatform": "There is no Platform",
         "noPlatformDesc": "Please make some transactions on investment platforms first.",
@@ -322,6 +323,7 @@
     "seeAllTransactions": "See all Transactions",
     "pendingTransactions": "Pending Transactions",
     "confirmedTransactions": "Confirmed Transactions",
+    "recentTransactions": "Recent Transactions",
     "showMore": "Show More",
     "invest": "Invest",
     "openOrders": "Open Orders",
