+++ src/components/HeaderDropdown/CurrencyDropdown.js
@@ -1,93 +1,13 @@
 import { useEffect } from "react";
-import styled from "styled-components";
-import { Button as BootstrapButton, Row, Col } from "react-bootstrap";
-import { Link, withRouter } from "react-router-dom";
+import { Row, Col } from "react-bootstrap";
+import { withRouter } from "react-router-dom";
 import { useDispatch, useSelector } from "react-redux";
 import { useTranslation } from "react-i18next";
-import SVG from 'react-inlinesvg';
-import Img from "../UI/Img";
+import SVG from "react-inlinesvg";
 import { rates } from "../../constants";
 import * as actions from "../../state/currency/actions";
 import FlagIcon from "../../assets/images/flag.svg";
-
-const Wrapper = styled.div`
-	position: relative;
-	padding: 14px 0;
-
-	&:hover .header-dropdown {
-		opacity: 1;
-		visibility: visible;
-		transform: rotateX(0deg) scale(1);
-	}
-`;
-
-const Item = styled.div`
-	display: flex;
-	align-items: center;
-`;
-
-const DropDown = styled.div`
-	position: absolute;
-	top: 100%;
-	right: 0;
-	background-color: ${({ theme }) => theme.modalBG};
-	border-radius: 0.75rem;
-	box-shadow: -1px 11px 43px rgba(0, 0, 0, 0.12);
-	padding: 20px;
-	width: auto;
-	min-width: 260px;
-
-	opacity: 0;
-	visibility: hidden;
-	transform-style: preserve-3d;
-	transform: rotateX(-40deg) scale(0.8);
-	transform-origin: top center;
-	transition: 0.4s ease all;
-	z-index: 99999;
-`;
-
-const DropDownItem = styled(Link)`
-	display: flex;
-	align-items: center;
-	position: relative;
-	outline: none;
-	text-decoration: none;
-	white-space: nowrap;
-	color: ${({ theme }) => theme.text1};
-
-	&:not(:last-child) {
-		margin-bottom: 20px;
-	}
-
-	&:hover,
-	&:focus,
-	&:active {
-		color: ${({ theme }) => theme.text2};
-		box-shadow: none;
-		outline: none;
-		text-decoration: none;
-	}
-`;
-
-const CurrencyLogo = styled(Img)`
-	width: 24px;
-	height: 24px;
-	margin-right: 20px;
-	border-radius: 24px;
-	border: 1px solid #a890fe;
-	min-width: 24px;
-`;
-
-const FlagButton = styled(BootstrapButton)`
-	color: ${({ theme }) => theme.primary};
-	margin-right: 12px;
-`;
-
-const Title = styled.span`
-	font-weight: 500;
-	font-size: 1rem;
-	color: ${({ theme }) => theme.text2};
-`;
+import { Wrapper, Item, IconButton, DropDown, DropDownItem, CurrencyLogo, Title2 } from "./styles";
 
 const HeaderDropdown = ({ items, title, ...props }) => {
 	const selectedCurrency = useSelector((state) => state.currency.selected);
@@ -106,9 +26,9 @@ const HeaderDropdown = ({ items, title, ...props }) => {
 	return (
 		<Wrapper>
 			<Item>
-				<FlagButton variant={"link"} className={"py-0 px-0"}>
+				<IconButton variant={"link"} className={"py-0 px-0"}>
 					<SVG src={FlagIcon} />
-				</FlagButton>
+				</IconButton>
 			</Item>
 			<DropDown className={"header-dropdown"}>
 				<Row>
@@ -119,7 +39,7 @@ const HeaderDropdown = ({ items, title, ...props }) => {
 								key={`currency-${currency}`}
 							>
 								<CurrencyLogo src={rates[currency].image} alt={currency} />
-								<Title>{currency}</Title>
+								<Title2>{currency}</Title2>
 							</DropDownItem>
 						</Col>
 					))}
