+++ src/components/TransactionsList/AllTransactionsList.js
@@ -1,0 +1,10 @@
+import { useAllTransactions } from "../../state/transactions/hooks";
+import TransactionsList from "./index";
+
+const AllTransactionsList = () => {
+	const transactions = useAllTransactions();
+
+	return <TransactionsList transactions={transactions} />;
+};
+
+export default AllTransactionsList;
